chore: Keil-v5 TM4C123 - uart printf - valvano Vol 1 Program 2.1
This commit contains the project configuration.
This is the program from Prof. Valvano. It shows how the 'printf'
statement is used together with the UART0-Peripheral in Port A of the
TM4C123 microcontroller.The Program can be downloaded from his website
for Keil-v4:
'users.ece.utexas.edu' (Serial Interfaces - Printf_UART_4C123.zip).
It has a slightly difference compared to the Program 2.1 from
Volume 1 - Embedded Systems 'Introduction to ARM Cortex-M
Microcontrollers' book. This program is ported to Keil v5 because of
the integrated Software Version Control System (SVCS) in the IDE. This
program is a demonstration of how printf outputs to 'Real-Term' are
printed. The baudrate in RealTerm is set to 115200 Bps.

This commit and projects solves the problem of having a working
program using the UART0 perihperal hardwaren together with printf
statements in the software for displaying data to 'Real-Term' for the
TM4C123 microcontroller in Keil-v5.